requirements.txt
# ============================================
# SUPER FOOD - Dependências do Sistema
# Versão: 2.0 (SQLAlchemy Multi-Tenant)
# ============================================

# ==================== CORE ====================

# Banco de Dados - SQLAlchemy ORM
sqlalchemy>=2.0.25

# Migrations (para produção)
alembic>=1.13.1

# Variáveis de Ambiente
python-dotenv>=1.0.1

# ==================== FRONTEND ====================

# Streamlit - Interface Web
streamlit>=1.40.2

# Data manipulation
pandas>=2.2.3

# Mapas interativos
pydeck>=0.9.1

# Gráficos
plotly>=5.24.1

# ==================== BACKEND API (OPCIONAL) ====================

# FastAPI + Uvicorn
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Autenticação JWT
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validação
pydantic[email]>=2.10.3

# Upload de arquivos
python-multipart>=0.0.20

# WebSockets (realtime)
websockets>=14.1

# ==================== INTEGRAÇÕES ====================

# HTTP Client (Mapbox, APIs externas)
requests>=2.32.3

# Cálculo de Distâncias
haversine>=2.8.1

# ==================== PRODUÇÃO ====================

# PostgreSQL Driver (quando migrar para nuvem)
psycopg2-binary>=2.9.9

# ==================== DESENVOLVIMENTO (OPCIONAL) ====================

# pytest>=8.3.4
# black>=24.10.0
# flake8>=7.1.1
# mypy>=1.7.0

# ============================================
# NOTAS:
# - Biblioteca sqlite3 já vem no Python
# - Biblioteca hashlib já vem no Python
# - Para produção, descomente psycopg2-binary
# ============================================